created a constants folder and moved the uni data
